requirements.txt
Flask==2.0.3
Pillow==9.0.0
PyYAML==6.0
moviepy==1.0.3